+++ packages/deploy-tooling/src/cli/config.ts
@@ -136,6 +136,26 @@ function mergeTarget(baseTarget: AbapTarget, options: CliOptions) {
 }
 
 /**
+ * Merge CLI and environment credentials.
+ *
+ * @param taskConfig - base configuration from the file
+ * @param options - CLI options
+ * @returns merged credentials
+ */
+function mergeCredentials(taskConfig: AbapDeployConfig, options: CliOptions) {
+    let credentials = taskConfig.credentials;
+    // Support CLI params and|or dotenv file
+    if (options.username || process.env.SERVICE_USERNAME) {
+        credentials = {
+            ...(credentials ?? {}),
+            username: options.username ?? process.env.SERVICE_USERNAME ?? '',
+            password: options.password ?? process.env.SERVICE_PASSWORD ?? ''
+        };
+    }
+    return credentials;
+}
+
+/**
  * Merge the configuration from the ui5*.yaml with CLI options.
  *
  * @param taskConfig - base configuration from the file
@@ -150,7 +170,7 @@ export async function mergeConfig(taskConfig: AbapDeployConfig, options: CliOpti
         transport: options.transport ?? taskConfig.app?.transport
     };
     const target = mergeTarget(taskConfig.target, options);
-    const config: AbapDeployConfig = { app, target, credentials: taskConfig.credentials };
+    const config: AbapDeployConfig = { app, target, credentials: mergeCredentials(taskConfig, options) };
     config.test = mergeFlag(options.test, taskConfig.test);
     config.safe = mergeFlag(options.safe, taskConfig.safe);
     config.keep = mergeFlag(options.keep, taskConfig.keep);